+++ malwarecage/cli/web.py
@@ -57,7 +57,11 @@ def npm_build_web(target_dir):
         if subprocess.call("npm run build", shell=True, cwd=context_dirname):
             raise BuildWebError(f"'npm run build' command failed")
 
-        click.echo(f"[*] Collecting artifacts", err=True)
+        if os.path.exists(target_dir):
+            click.echo(f"[*] Target {target_dir} exists, removing...", err=True)
+            shutil.rmtree(target_dir)
+
+        click.echo(f"[*] Collecting artifacts to {target_dir}", err=True)
         shutil.move(
             os.path.join(context_dirname, "build"),
             target_dir
